requirements.txt
pandas
numpy
langchain
langchain-openai
python-dotenv
langchainhub